--- src/components/molecules/NavBar.tsx
@@ -1,10 +1,0 @@
-import Link from 'next/link';
-
-export default function NavBar(): JSX.Element {
-  return (
-    <div>
-      <Link href="/">HOME</Link>
-      <Link href="/login">로그인페이지로</Link>
-    </div>
-  );
-}
