+++ src/components/organisms/common/Header.tsx
@@ -1,13 +1,8 @@
-import NavBar from '@/components/molecules/NavBar';
-
-import HeaderNav from '@/components/molecules/header-elements/HeaderNav';
 import HeaderNavigate from '@/components/molecules/header-elements/HeaderNavigate';
 
 const Header = () => {
   return (
     <div>
-      <NavBar />
-      {/* <HeaderNav /> */}
       <HeaderNavigate />
     </div>
   );
