+++ src/components/organisms/common/floatingbar/FloatingBar.tsx
@@ -1,7 +1,7 @@
 import UserIcon from '@/components/common/UserIcon';
 import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
-import { styled } from 'styled-components';
+import * as S from './styled';
 import {
   MdOutlineSearch,
   MdKeyboardArrowUp,
@@ -11,36 +11,6 @@ import { IoMdChatbubbles } from 'react-icons/io';
 import Link from 'next/link';
 import ChatModal from '@/components/molecules/chat-modal/ChatModal';
 import useModal from '@/hooks/useModal';
-
-interface FloatingBoxProps {
-  position: number;
-}
-
-const FloatingBox = styled.div<FloatingBoxProps>`
-  margin: 10px;
-  position: absolute;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  top: ${(props) => props.position}px;
-  right: 50px;
-  /* border: 3px solid #9d9d9d; */
-`;
-const FriendSearchIcon = styled.div`
-  font-size: 30px;
-  color: #91c8e4;
-`;
-
-const ChatIcon = styled.div`
-  font-size: 30px;
-  color: #91c8e4;
-  cursor: pointer;
-`;
-const ScrollButotn = styled.div`
-  font-size: 30px;
-  cursor: pointer;
-  opacity: 0.7;
-`;
 
 const scrollToTop = () => {
   window.scrollTo({ top: 0, behavior: 'smooth' });
@@ -82,32 +52,32 @@ const FloatingBar = () => {
   const { isOpenModal: chatState, handleModal: chatHandle } = useModal();
 
   return (
-    <FloatingBox position={floatingPosition}>
+    <S.FloatingBox position={floatingPosition}>
       <a>
-        <ScrollButotn>
+        <S.ScrollButotn>
           <MdKeyboardArrowUp onClick={scrollToTop} />
-        </ScrollButotn>
+        </S.ScrollButotn>
       </a>
       <Link legacyBehavior href="/search/SearchFriends">
         <a>
-          <FriendSearchIcon>
+          <S.FriendSearchIcon>
             <MdOutlineSearch />
-          </FriendSearchIcon>
+          </S.FriendSearchIcon>
         </a>
       </Link>
       <UserIcon />
       <a>
-        <ChatIcon onClick={chatHandle}>
+        <S.ChatIcon onClick={chatHandle}>
           <IoMdChatbubbles />
-        </ChatIcon>
+        </S.ChatIcon>
         {chatState && <ChatModal show={chatState} hide={chatHandle} />}
       </a>
       <a>
-        <ScrollButotn>
+        <S.ScrollButotn>
           <MdKeyboardArrowDown onClick={scrollToBottom} />
-        </ScrollButotn>
+        </S.ScrollButotn>
       </a>
-    </FloatingBox>
+    </S.FloatingBox>
   );
 };
