+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import Footer from '@/components/organisms/common/Footer';
 import Header from '@/components/organisms/common/Header';
 import type { AppProps } from 'next/app';
 import { RecoilRoot } from 'recoil';
-import FloatingBar from '@/components/organisms/common/FloatingBar';
+import FloatingBar from '@/components/organisms/common/floatingbar/FloatingBar';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
